+++ src/main/kotlin/me/ezra_home/retail_software_solution/organizations/business/unitvalue/api/UnitValueResponseDto.kt
@@ -1,6 +1,7 @@
 package me.ezra_home.retail_software_solution.organizations.business.unitvalue.api
 
 import java.io.Serializable
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -11,7 +12,7 @@ data class UnitValueResponseDto(
     val description: String?,
     val baseUnit: UUID?,
     val baseUnitName: String?,
-    val conversionFactor: Double?,
+    val conversionFactor: BigDecimal?,
     val createdBy: String?,
     val createdOn: OffsetDateTime?,
     val unitGroupId: UUID,
